fix: UBSan violations and compiler warnings in game code
Fixes a set of undefined behaviour and signed/unsigned mismatch issues
found by UBSan while playing through the game, plus two warning fixes.

Unsigned-shift UB (signed left shift of 1):
- soh/include/z64actor.h: ACTOR_FLAG_* macros (1 << N) -> (1u << N)
- soh/include/z64player.h: PLAYER_STATE* macros same
- soh/src/code/z_actor.c: Flags_GetSwitch / Set / Unset use (1u << flag)
- soh/src/code/z_camera.c, z_quake.c: explicit s16 casts on shift inputs

Pointer/size mismatch UB:
- soh/include/alignment.h: ALIGN* macros cast to uintptr_t before bitwise
- soh/src/code/TwoHeadArena.c: mask variables as uintptr_t
- soh/src/libultra/gu/guLookAtHilite.c: FTOFRAC8 saturates to INT8 range

Signed/unsigned comparison warnings:
- soh/include/functions.h: Flags_GetSwitch return u32, flag u8
- soh/include/z64.h: add struct tag names (TitleContext, OpeningContext)
- soh/soh/ActorDB.cpp, util.cpp, gameplaystats.cpp: size_t comparisons
- soh/soh/SohGui/UIWidgets.cpp: loop index size_t
- soh/soh/Enhancements/randomizer/option.cpp, dungeon.cpp: size_t / init order

Other fixes:
- soh/src/code/code_800E4FE0.c: explicit cast before shift
- soh/src/code/code_800EC960.c: timing vars as u64 (were u32, overflowed)
- soh/src/code/code_800F9280.c: explicit uint8 wraparound for ring index
- soh/src/code/padmgr.c: buttonDiff as u32 (bitwise op on signed)
- soh/src/code/sys_math_atan.c: guard against y==0, use lrintf
- soh/soh/ResourceManagerHelpers.cpp, stubs.c: [[noreturn]] attribute
- soh/soh/Enhancements/randomizer/ShuffleBeehives.cpp: || -> && (logic bug)
- soh/soh/Enhancements/randomizer/SeedContext.h: InitStaticData -> static
- All ovl_* actor files: various type/cast fixes found by UBSan

+++ soh/include/functions.h
@@ -373,9 +373,9 @@ void Actor_SetFeetPos(Actor* actor, s32 limbIndex, s32 leftFootIndex, Vec3f* lef
                       Vec3f* rightFootPos);
 void func_8002BE04(PlayState* play, Vec3f* arg1, Vec3f* arg2, f32* arg3);
 void func_8002C124(TargetContext* targetCtx, PlayState* play);
-s32 Flags_GetSwitch(PlayState* play, s32 flag);
-void Flags_SetSwitch(PlayState* play, s32 flag);
-void Flags_UnsetSwitch(PlayState* play, s32 flag);
+u32 Flags_GetSwitch(PlayState* play, u8 flag);
+void Flags_SetSwitch(PlayState* play, u8 flag);
+void Flags_UnsetSwitch(PlayState* play, u8 flag);
 s32 Flags_GetUnknown(PlayState* play, s32 flag);
 void Flags_SetUnknown(PlayState* play, s32 flag);
 void Flags_UnsetUnknown(PlayState* play, s32 flag);
